refactor CheckIfc to use based on given IFC file domain
