+++ bim2sim/assets/templates/check_ifc/summary_template
@@ -26,7 +26,7 @@
     <tbody>
         <tr>
             <th colspan=2 class="tg-7btt">
-            <a href="${plugin_name}_error_summary_inst.html">Errors in IFC</a>
+            <a href="${base_name}_error_summary_inst.html">Errors in IFC</a>
             </th>
         </tr>
         <tr>
@@ -49,7 +49,7 @@
         </tr>
         <tr>
             <th  colspan=2 class="tg-7btt"><a href=
-                "${plugin_name}_error_summary_prop.html">Missing Attributes for
+                "${base_name}_error_summary_prop.html">Missing Attributes for
                 ${plugin_name}</a>
             </th>
         </tr>
