+++ bim2sim/plugins/PluginEnergyPlus/bim2sim_energyplus/__init__.py
@@ -18,7 +18,7 @@ class PluginEnergyPlus(Plugin):
     elements = {*bps_elements.items, Material} - {bps_elements.Plate}
     default_tasks = [
         common.LoadIFC,
-        bps.CheckIfcBPS,
+        common.CheckIfc,
         common.CreateElements,
         bps.CreateSpaceBoundaries,
         bps.Prepare,
