+++ bim2sim/plugins/PluginHKESim/bim2sim_hkesim/__init__.py
@@ -26,7 +26,7 @@ class PluginHKESim(Plugin):
     elements = {*hvac_elements.items, Material}
     default_tasks = [
         common.LoadIFC,
-        hvac.CheckIfcHVAC,
+        hvac.CheckIfc,
         common.CreateElements,
         hvac.ConnectElements,
         hvac.MakeGraph,
