+++ bim2sim/plugins/PluginSpawnOfEP/bim2sim_spawn/__init__.py
@@ -32,7 +32,7 @@ class PluginCoSimModelica(Plugin):
 
     default_tasks = [
         common.LoadIFC,
-        hvac.CheckIfcHVAC,
+        hvac.CheckIfc,
         common.CreateElements,
         hvac.ConnectElements,
         hvac.MakeGraph,
