+++ bim2sim/task/bps/__init__.py
@@ -1,5 +1,4 @@
 from .bind_tz import CombineThermalZones
-from .check_ifc import CheckIfcBPS
 from .disaggr_creation import DisaggregationCreation
 from .enrich_material import EnrichMaterial
 from .enrich_use_cond import EnrichUseConditions
