+++ bim2sim/task/common/__init__.py
@@ -1,4 +1,4 @@
 from .bind_storeys import BindStoreys
-from .common import Reset, Quit, LoadIFC, CreateElements
+from .common_tasks import Reset, Quit, LoadIFC, CreateElements, CheckIfc
 from .weather import Weather
