+++ bim2sim/task/hvac/__init__.py
@@ -1,4 +1,3 @@
-from .check_ifc import CheckIfcHVAC
 from .dead_ends import DeadEnds
 from .expansiontanks import ExpansionTanks
 from .fixports import FixPorts
